+++ modules/tolyui_navigation/lib/src/rail_menu_tree/model/menu_tree_state.dart
@@ -6,6 +6,7 @@ class MenuTreeMeta {
   final List<String> expandMenus;
   final MenuNode? activeMenu;
   final MenuNode root;
+
    List<MenuNode> get items=>root.children;
 
   MenuTreeMeta({