+++ modules/tolyui_feedback/CHANGELOG.md
@@ -6,3 +6,8 @@
 ## 0.3.1
 
 * Support Global Message Notification by [tolyui_message](https://pub.dev/packages/tolyui_message)
+
+## 0.3.3
+
+* DecorationConfig#BubbleMeta config Bubble △ height and angle
+* [menuOffsetCalculator] to translate pop panel