+++ modules/tolyui_navigation/CHANGELOG.md
@@ -3,4 +3,7 @@
 * Support Widget: [TolyRailMenuBar]
 
 ## 0.0.2
-* Support Widget: [TolyRailMenuBar#cellStyle]
+* Support style: [TolyRailMenuBar#cellStyle]
+
+## 0.0.5
+* Support Widget: [TolyRailMenuTree]